Add AUR install instructions, fix indentation on Deb, RPM or Portable Section
Add installation instructions for installing from AUR and connecting with the Browser extensions. (AUR installs to `/opt/JabRef/` which doesn't work with the linked .json messaging host files since the path in them has a lowercase `jabref`).
Also fixed the indentation on the Deb, RPM or Portable Section to have a consistent use of 3 Spaces as indent.